Removiendo estructura antigua de documentación. Completando flujo de compilación y limpieza de fuentes.

+++ tribus/common/clean.py
@@ -2,7 +2,8 @@ import os
 from distutils.cmd import Command
 from distutils.command.clean import clean as base_clean
 
-from tribus.config.pkg import png_file_list, mo_file_list
+from tribus.config.base import BASEDIR, DOCDIR
+from tribus.common.utils import get_path, find_files, find_dirs
 from tribus.common.logger import get_logger
 
 log = get_logger()
@@ -19,13 +20,21 @@ class clean_img(Command):
         pass
 
     def run(self):
-        for png_file in png_file_list:
-            os.remove(png_file)
-            log.debug("[%s.%s] Removing \"%s\"." % (__name__, self.__class__.__name__, png_file))
+        for png_file in find_files(path=BASEDIR, pattern='*.png'):
+            if os.path.isfile(png_file):
+                try:
+                    os.remove(png_file)
+                    log.debug("[%s.%s] Removing \"%s\"." % (__name__,
+                                                            self.__class__.__name__,
+                                                            png_file))
+                except Exception, e:
+                    print e
 
 
 class clean_html(Command):
     description = 'Compile .po files into .mo files'
+    user_options = []
+
     def initialize_options(self):
         pass
  
@@ -33,11 +42,26 @@ class clean_html(Command):
         pass
  
     def run(self):
-        print 'Cleaning html'
+        for html_file in find_files(path=get_path([DOCDIR, 'html']), pattern='*.*'):
+            if os.path.isfile(html_file):
+                try:
+                    os.remove(html_file)
+                    log.debug("[%s.%s] Removing \"%s\"." % (__name__, self.__class__.__name__, html_file))
+                except Exception, e:
+                    print e
+
+        for html_dir in reversed(find_dirs(path=get_path([DOCDIR, 'html']))):
+            try:
+                os.rmdir(html_dir)
+                log.debug("[%s.%s] Removing \"%s\"." % (__name__, self.__class__.__name__, html_dir))
+            except Exception, e:
+                print e
 
 
 class clean_man(Command):
     description = 'Compile .po files into .mo files'
+    user_options = []
+
     def initialize_options(self):
         pass
  
@@ -45,11 +69,19 @@ class clean_man(Command):
         pass
  
     def run(self):
-        print 'Cleaning man'
+        man_file = get_path([DOCDIR, 'man', 'tribus.1'])
+        if os.path.isfile(man_file):
+            try:
+                os.remove(man_file)
+                log.debug("[%s.%s] Removing \"%s\"." % (__name__, self.__class__.__name__, man_file))
+            except Exception, e:
+                print e
 
 
 class clean_mo(Command):
     description = 'Compile .po files into .mo files'
+    user_options = []
+
     def initialize_options(self):
         pass
  
@@ -57,9 +89,13 @@ class clean_mo(Command):
         pass
  
     def run(self):
-        for mo_file in mo_file_list:
-            os.remove(mo_file)
-            log.debug("[%s.%s] Removing \"%s\"." % (__name__, self.__class__.__name__, mo_file))
+        for mo_file in find_files(path=BASEDIR, pattern='*.mo'):
+            if os.path.isfile(mo_file):
+                try:
+                    os.remove(mo_file)
+                    log.debug("[%s.%s] Removing \"%s\"." % (__name__, self.__class__.__name__, mo_file))
+                except Exception, e:
+                    print e
 
 class clean(base_clean):
     def run(self):
